--story=872097271 rds 注册触发事件规则

+++ src/backend/ci/ext/tencent/rds/api-rds-tencent/src/main/kotlin/com/tencent/devops/rds/pojo/RdsChartPipelineInfo.kt
@@ -35,7 +35,7 @@ data class RdsChartPipelineInfo(
     @ApiModelProperty("对应流水线ID")
     val pipelineId: String,
     @ApiModelProperty("自定义ID")
-    val productId: Int,
+    val productId: Long,
     @ApiModelProperty("所在Chart相对路径")
     val filePath: String,
     @ApiModelProperty("CLI解析后的YAML内容")
