+++ src/backend/ci/ext/tencent/rds/biz-rds-tencent/src/main/kotlin/com/tencent/devops/rds/chart/StreamConverter.kt
@@ -65,7 +65,7 @@ class StreamConverter @Autowired constructor(
 
     fun buildModel(
         userId: String,
-        productId: Int,
+        productId: Long,
         projectId: String,
         cachePath: String,
         pipelineFile: File
